Split Android cards feature packages
