+++ apps/android/feature/cards/src/main/java/com/flashcardsopensourceapp/feature/cards/editor/CardEditorUiState.kt
@@ -1,17 +1,7 @@
-package com.flashcardsopensourceapp.feature.cards
+package com.flashcardsopensourceapp.feature.cards.editor
 
-import com.flashcardsopensourceapp.data.local.model.cards.CardFilter
-import com.flashcardsopensourceapp.data.local.model.cards.CardSummary
 import com.flashcardsopensourceapp.data.local.model.scheduling.EffortLevel
 import com.flashcardsopensourceapp.data.local.model.workspace.WorkspaceTagSummary
-
-data class CardsUiState(
-    val isLoading: Boolean,
-    val searchQuery: String,
-    val activeFilter: CardFilter,
-    val availableTagSuggestions: List<WorkspaceTagSummary>,
-    val cards: List<CardSummary>
-)
 
 data class CardEditorUiState(
     val isLoading: Boolean,
