+++ apps/android/feature/cards/src/main/java/com/flashcardsopensourceapp/feature/cards/editor/CardEditorViewModel.kt
@@ -1,209 +1,24 @@
-package com.flashcardsopensourceapp.feature.cards
+package com.flashcardsopensourceapp.feature.cards.editor
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import androidx.lifecycle.viewmodel.CreationExtras
-import androidx.lifecycle.viewmodel.initializer
-import androidx.lifecycle.viewmodel.viewModelFactory
-import com.flashcardsopensourceapp.core.ui.TransientMessageController
-import com.flashcardsopensourceapp.core.ui.VisibleAppScreen
-import com.flashcardsopensourceapp.core.ui.VisibleAppScreenRepository
 import com.flashcardsopensourceapp.data.local.model.cards.CardDraft
-import com.flashcardsopensourceapp.data.local.model.cards.CardFilter
 import com.flashcardsopensourceapp.data.local.model.cards.CardSummary
+import com.flashcardsopensourceapp.data.local.model.cards.normalizeTags
 import com.flashcardsopensourceapp.data.local.model.scheduling.EffortLevel
 import com.flashcardsopensourceapp.data.local.model.workspace.WorkspaceTagSummary
-import com.flashcardsopensourceapp.data.local.model.cards.normalizeTags
-import com.flashcardsopensourceapp.data.local.repository.sync.AutoSyncCompletion
-import com.flashcardsopensourceapp.data.local.repository.sync.AutoSyncEvent
-import com.flashcardsopensourceapp.data.local.repository.sync.AutoSyncEventRepository
-import com.flashcardsopensourceapp.data.local.repository.sync.AutoSyncOutcome
-import com.flashcardsopensourceapp.data.local.repository.sync.AutoSyncRequest
 import com.flashcardsopensourceapp.data.local.repository.CardsRepository
 import com.flashcardsopensourceapp.data.local.repository.WorkspaceRepository
-import kotlinx.coroutines.ExperimentalCoroutinesApi
+import com.flashcardsopensourceapp.feature.cards.CardsTextProvider
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-
-@OptIn(ExperimentalCoroutinesApi::class)
-class CardsViewModel(
-    private val cardsRepository: CardsRepository,
-    private val autoSyncEventRepository: AutoSyncEventRepository,
-    private val messageController: TransientMessageController,
-    visibleAppScreenRepository: VisibleAppScreenRepository,
-    workspaceRepository: WorkspaceRepository,
-    private val textProvider: CardsTextProvider
-) : ViewModel() {
-    private val searchQuery = MutableStateFlow(value = "")
-    private val activeFilter = MutableStateFlow(
-        value = CardFilter(
-            tags = emptyList(),
-            effort = emptyList()
-        )
-    )
-
-    private val cardsFlow = combine(
-        searchQuery,
-        activeFilter
-    ) { query, filter ->
-        query to filter
-    }.flatMapLatest { (query, filter) ->
-        cardsRepository.observeCards(
-            searchQuery = query,
-            filter = filter
-        )
-    }
-    private val visibleAppScreenState = visibleAppScreenRepository.observeVisibleAppScreen().stateIn(
-        scope = viewModelScope,
-        started = SharingStarted.WhileSubscribed(stopTimeoutMillis = 5_000L),
-        initialValue = VisibleAppScreen.OTHER
-    )
-    private var pendingAutoSyncRequestId: String? = null
-    private var cardsSignatureAtAutoSyncStart: CardsVisibleSignature? = null
-    private var lastVisibleAutoSyncChangeSignature: CardsVisibleSignature? = null
-
-    val uiState: StateFlow<CardsUiState> = combine(
-        cardsFlow,
-        workspaceRepository.observeWorkspaceTagsSummary(),
-        searchQuery,
-        activeFilter
-    ) { cards, tagsSummary, query, filter ->
-        CardsUiState(
-            isLoading = false,
-            searchQuery = query,
-            activeFilter = filter,
-            availableTagSuggestions = tagsSummary.tags,
-            cards = cards
-        )
-    }.stateIn(
-        scope = viewModelScope,
-        started = SharingStarted.WhileSubscribed(stopTimeoutMillis = 5_000L),
-        initialValue = CardsUiState(
-            isLoading = true,
-            searchQuery = "",
-            activeFilter = CardFilter(tags = emptyList(), effort = emptyList()),
-            availableTagSuggestions = emptyList(),
-            cards = emptyList()
-        )
-    )
-
-    init {
-        observeAutoSyncDrivenCardsChanges()
-    }
-
-    fun updateSearchQuery(query: String) {
-        searchQuery.value = query
-    }
-
-    fun applyFilter(filter: CardFilter) {
-        activeFilter.value = filter
-    }
-
-    fun clearFilter() {
-        activeFilter.value = CardFilter(
-            tags = emptyList(),
-            effort = emptyList()
-        )
-    }
-
-    private fun observeAutoSyncDrivenCardsChanges() {
-        viewModelScope.launch {
-            autoSyncEventRepository.observeAutoSyncEvents().collect { event ->
-                when (event) {
-                    is AutoSyncEvent.Requested -> {
-                        handleAutoSyncRequested(request = event.request)
-                    }
-
-                    is AutoSyncEvent.Completed -> {
-                        handleAutoSyncCompleted(completion = event.completion)
-                    }
-                }
-            }
-        }
-    }
-
-    private fun handleAutoSyncRequested(request: AutoSyncRequest) {
-        if (request.allowsVisibleChangeMessage.not()) {
-            return
-        }
-        if (visibleAppScreenState.value != VisibleAppScreen.CARDS) {
-            return
-        }
-
-        pendingAutoSyncRequestId = request.requestId
-        cardsSignatureAtAutoSyncStart = buildCardsVisibleSignature(uiState = uiState.value)
-    }
-
-    private fun handleAutoSyncCompleted(completion: AutoSyncCompletion) {
-        if (completion.request.requestId != pendingAutoSyncRequestId) {
-            return
-        }
-
-        pendingAutoSyncRequestId = null
-        val cardsSignatureBeforeSync = cardsSignatureAtAutoSyncStart
-        cardsSignatureAtAutoSyncStart = null
-
-        if (completion.outcome !is AutoSyncOutcome.Succeeded) {
-            return
-        }
-        if (completion.request.allowsVisibleChangeMessage.not()) {
-            return
-        }
-        if (visibleAppScreenState.value != VisibleAppScreen.CARDS) {
-            return
-        }
-
-        val currentCardsSignature = buildCardsVisibleSignature(uiState = uiState.value)
-        if (cardsSignatureBeforeSync == null || cardsSignatureBeforeSync == currentCardsSignature) {
-            return
-        }
-        if (currentCardsSignature == lastVisibleAutoSyncChangeSignature) {
-            return
-        }
-
-        lastVisibleAutoSyncChangeSignature = currentCardsSignature
-        messageController.showMessage(message = textProvider.cardsUpdatedOnAnotherDeviceMessage)
-    }
-}
-
-private data class VisibleCardSignature(
-    val cardId: String,
-    val frontText: String,
-    val effortLevel: EffortLevel,
-    val tags: List<String>,
-    val dueAtMillis: Long?
-)
-
-private data class CardsVisibleSignature(
-    val searchQuery: String,
-    val activeFilter: CardFilter,
-    val cards: List<VisibleCardSignature>
-)
-
-private fun buildCardsVisibleSignature(uiState: CardsUiState): CardsVisibleSignature {
-    return CardsVisibleSignature(
-        searchQuery = uiState.searchQuery,
-        activeFilter = uiState.activeFilter,
-        cards = uiState.cards.map { card ->
-            VisibleCardSignature(
-                cardId = card.cardId,
-                frontText = card.frontText,
-                effortLevel = card.effortLevel,
-                tags = card.tags,
-                dueAtMillis = card.dueAtMillis
-            )
-        }
-    )
-}
 
 private data class CardEditorDraftState(
     val frontText: String,
@@ -468,46 +283,6 @@ class CardEditorViewModel(
     }
 }
 
-fun createCardsViewModelFactory(
-    cardsRepository: CardsRepository,
-    workspaceRepository: WorkspaceRepository,
-    autoSyncEventRepository: AutoSyncEventRepository,
-    messageController: TransientMessageController,
-    visibleAppScreenRepository: VisibleAppScreenRepository
-): ViewModelProvider.Factory {
-    return viewModelFactory {
-        initializer {
-            val application = requireApplication()
-            CardsViewModel(
-                cardsRepository = cardsRepository,
-                autoSyncEventRepository = autoSyncEventRepository,
-                messageController = messageController,
-                visibleAppScreenRepository = visibleAppScreenRepository,
-                workspaceRepository = workspaceRepository,
-                textProvider = cardsTextProvider(context = application)
-            )
-        }
-    }
-}
-
-fun createCardEditorViewModelFactory(
-    cardsRepository: CardsRepository,
-    workspaceRepository: WorkspaceRepository,
-    editingCardId: String?
-): ViewModelProvider.Factory {
-    return viewModelFactory {
-        initializer {
-            val application = requireApplication()
-            CardEditorViewModel(
-                cardsRepository = cardsRepository,
-                workspaceRepository = workspaceRepository,
-                editingCardId = editingCardId,
-                textProvider = cardsTextProvider(context = application)
-            )
-        }
-    }
-}
-
 private data class CardEditorValidationResult(
     val isValid: Boolean,
     val frontTextErrorMessage: String,
@@ -539,10 +314,6 @@ private fun validateCardEditorInput(
     )
 }
 
-private fun CreationExtras.requireApplication(): android.app.Application {
-    return checkNotNull(this[ViewModelProvider.AndroidViewModelFactory.APPLICATION_KEY])
-}
-
 private fun toggleTagSelection(selectedTags: List<String>, tag: String): List<String> {
     if (selectedTags.contains(tag)) {
         return selectedTags.filter { value ->
@@ -551,22 +322,4 @@ private fun toggleTagSelection(selectedTags: List<String>, tag: String): List<St
     }
 
     return selectedTags + tag
-}
-
-internal fun buildCardEditorDraft(
-    frontText: String,
-    backText: String,
-    selectedTags: List<String>,
-    effortLevel: EffortLevel,
-    referenceTags: List<String>
-): CardDraft {
-    return CardDraft(
-        frontText = frontText.trim(),
-        backText = backText.trim(),
-        tags = normalizeTags(
-            values = selectedTags,
-            referenceTags = referenceTags
-        ),
-        effortLevel = effortLevel
-    )
 }
